dhp-store/ai-service/requirements.txt
fastapi
uvicorn
mysql-connector-python
pandas
scikit-learn
python-dotenv
requests
google-generativeai